requirements.txt
requests
gspread
aiohttp
google-auth
google-auth-oauthlib
google-api-python-client
PyPDF2
customtkinter
python-dotenv
nest-asyncio
